Implementing an option to continue processing.
We often have apache access logs that contain unparsable or badly
formatted data.  This patch enables a new config option to continue on
processing when a bad record/line is encountered.

The default is to stop whenever unparsable lines are encountered and
this does not change.

If you add the following line to your yaml config, the parser will skip
unparsable lines so that it can continue processing.

`stop_on_invalid_record: false`

This patch also imports access to the Exec objects Logger and will emit
log messages whenever it ignores log lines.